Adds work in progress for Task view
