+++ src/views/Task/Select.jsx
@@ -1,0 +1,12 @@
+return (
+  <GradientSelect
+    value={value}
+    filled
+    options={options.map(text => ({
+      text,
+      callback: () => {
+        next(index, text);
+      },
+    }))}
+  />
+);
