+++ src/views/Task/styled.js
@@ -1,0 +1,27 @@
+import styled from 'styled-components';
+
+
+const SubmitWrap = styled.div`
+  padding-top: 20px;
+`;
+
+const FlexWrap = styled.div`
+  display: flex;
+`;
+
+const ButtonWrap = styled.div`
+  padding-right: 30px;
+`;
+
+
+export {
+  SubmitWrap,
+  FlexWrap,
+  ButtonWrap,
+};
+
+export default {
+  SubmitWrap,
+  FlexWrap,
+  ButtonWrap,
+};
